Use MergeSpecification for event tracking, and set mock event listener randomly into LuceneTestCase.newIndexConfig
