+++ lucene/test-framework/src/java/org/apache/lucene/index/MockIndexWriterEventListener.java
@@ -1,0 +1,37 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.lucene.index;
+
+/** Mock IndexWriterEventListener to verify invocation of event methods */
+public class MockIndexWriterEventListener implements IndexWriterEventListener {
+  private boolean beginMergeCalled = false;
+  private boolean endMergeCalled = false;
+
+  @Override
+  public void beginMergeOnFullFlush(MergePolicy.MergeSpecification merge) {
+    beginMergeCalled = true;
+  }
+
+  @Override
+  public void endMergeOnFullFlush(MergePolicy.MergeSpecification merge) {
+    endMergeCalled = true;
+  }
+
+  public boolean isEventsRecorded() {
+    return beginMergeCalled && endMergeCalled;
+  }
+}
